Disabled URL Direct Access For All Pages

+++ Back-end/se_project/shoes/views.py
@@ -4,12 +4,18 @@ from .form import *
 from random import randint, randrange
 from django.urls import reverse
 from django.http.response import JsonResponse
-from django.http import HttpResponseRedirect
+from django.http import HttpResponseRedirect, Http404
 from django.contrib.auth.forms import UserCreationForm, AuthenticationForm
 from django.contrib.auth.decorators import login_required
 from django.contrib import messages
 from django.contrib.auth import authenticate, login
 
+
+def get_referer(request):
+    referer = request.META.get('HTTP_REFERER')
+    if not referer:
+        return None
+    return referer
 
 def home(request):
     x = [None for i in range(4)]
@@ -26,21 +32,22 @@ def home(request):
 
 
 def search_shoes(request):
-    if request.method == "POST":
-        searched = request.POST['searched']
-        shoes_brands = Shoe.objects.filter(brand__contains=searched)
-        context = {
-            'searched': searched,
-            'shoes_brands': shoes_brands,
-            'page_name': 'Search Results',
-            'nav': True
-        }
-        return render(request, 'shoes/search_shoes.html', context)
-    else:
-        return render(request, 'shoes/search_shoes.html', {'page_name': 'Search Results', 'nav': True})
+    if not get_referer(request):
+        raise Http404
+    searched = request.POST['searched']
+    shoes_brands = Shoe.objects.filter(brand__contains=searched)
+    context = {
+        'searched': searched,
+        'shoes_brands': shoes_brands,
+        'page_name': 'Search Results',
+        'nav': True
+    }
+    return render(request, 'shoes/search_shoes.html', context)
 
 
 def item_page(request, myid):
+    if not get_referer(request):
+        raise Http404
     context = {
         'shoes': Shoe.objects.filter(id=myid),
         'page_name': 'Item Page',
@@ -50,6 +57,8 @@ def item_page(request, myid):
 
 
 def registration(request):
+    if not get_referer(request):
+        raise Http404
     if request.method == "POST":
         form = RegistrationForm(request.POST)
         if form.is_valid():
@@ -82,6 +91,8 @@ def women(request):
 
 
 def cart(request):
+    if not get_referer(request):
+        raise Http404
     total_price = 0
     context = {
         'shoes': Shoe.objects.all(),
@@ -97,26 +108,33 @@ def cart(request):
                     temp += j
             total_price += int(temp) * i.product_qty
     total_items = context['cart_items'].count()
-    context['no_items'] = total_items
+    context['no_items'] = int(total_items)
     context['total_price'] = total_price
     context['price_with_shipping'] = total_price + 40
-
     return render(request, 'shoes/shoppingcart.html', context)
 
 def checkout(request):
-    x = randint(10000, 99999)
-    if request.method == "POST":
+    if not get_referer(request):
+        raise Http404
+    if User_Order.objects.all():
         for i in User_Order.objects.all():
             i.delete()
+        x = randint(10000, 99999)
         context = {
-            'order_no': x,
-            'page_name': "Checkout"
+            'page_name': "Checkout",
+            'order_no': x
         }
-        return render(request, 'shoes/checkout.html', context)
     else:
-        messages.error(request, "Please Fill Your Cart Then Checkout :(")
-        return redirect('cart')
+        context = {
+            'page_name': "Checkout",
+            'Error': "Stop Refreshing :("
+        }
+    return render(request, 'shoes/checkout.html', context)
+    # messages.success(request, "Please Fill Your Cart Then Checkout :(")
+    # return redirect('cart')
 def add_To_Cart(request, xid):
+    if not get_referer(request):
+        raise Http404
     if User_Order.objects.filter(product_id=xid):
         check = User_Order.objects.get(product_id=xid)
         check.product_qty += 1
@@ -128,6 +146,8 @@ def add_To_Cart(request, xid):
     return HttpResponseRedirect(request.META.get('HTTP_REFERER', '/'))
 
 def remove_From_Cart(request, xid):
+    if not get_referer(request):
+        raise Http404
     if User_Order.objects.filter(product_id=xid):
         check = User_Order.objects.get(product_id=xid)
         check.delete()
